+++ shared/form-designer/components.d.ts
@@ -54,5 +54,7 @@ declare module 'vue' {
     QToolbar: typeof import('quasar')['QToolbar']
     QToolbarTitle: typeof import('quasar')['QToolbarTitle']
     QTr: typeof import('quasar')['QTr']
+    RouterLink: typeof import('vue-router')['RouterLink']
+    RouterView: typeof import('vue-router')['RouterView']
   }
 }