+++ shared/form-apis/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @herodotus-cloud/form-apis
 
+## 4.6.10
+
+### Patch Changes
+
+- 版本发布
+
+## 4.6.9
+
+### Patch Changes
+
+- vite 版本升级至 8.0.16
+
 ## 4.6.8
 
 ### Patch Changes