+++ shared/form-engine/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @herodotus-cloud/form-engine
 
+## 4.6.10
+
+### Patch Changes
+
+- 版本发布
+- Updated dependencies
+  - @herodotus-cloud/form-apis@4.6.10
+
+## 4.6.9
+
+### Patch Changes
+
+- vite 版本升级至 8.0.16
+- Updated dependencies
+  - @herodotus-cloud/form-apis@4.6.9
+
 ## 4.6.8
 
 ### Patch Changes